move the duplicate event subscriber as an inner class into the github duplicate monitor
